analysis-worker/requirements.txt
# Analysis worker dependencies
# Pin mediapipe to 0.10.14 - minimum required by OceanAI 1.0.0a48
# Later versions may have removed 'solutions' API
mediapipe==0.10.14
torchcodec
openai>=1.0.0
boto3
python-dotenv
ipython
opensmile
gradio

# Big Five / OceanAI
oceanai==1.0.0a48

# Scientific + ML utilities (kept here so local `uv pip install -r ...` matches Docker)
numpy
pandas
scikit-learn
librosa
opencv-python-headless
transformers
sentencepiece
# OceanAI/transformers ecosystem can be sensitive to protobuf versions.
protobuf==3.20.3

# Serverless HTTP wrapper (used by analysis-worker/Dockerfile runtime)
flask
gunicorn
